refactor: excluído o componente appBar e o modal foi repassado para a pages cars e users
