+++ frontend/src/pages/brands/index.tsx
@@ -31,6 +31,9 @@ const Brands = () => {
   return (
     <>
       <Header />
+
+
+      
       <Container>
         <ContainerTable>
           <Table>
